Expect uint64_t as output value
Just to prevent overflow answers

+++ aoc2020/include/days.h
@@ -3,13 +3,13 @@
 class Day01 : public Problem {
 public:
     Day01(const std::string& input) : Problem(input) {}
-    std::pair<bool, std::uint32_t> part1() override;
-    std::pair<bool, std::uint32_t> part2() override;
+    std::pair<bool, std::uint64_t> part1() override;
+    std::pair<bool, std::uint64_t> part2() override;
 };
 
 class Day02 : public Problem {
 public:
     Day02(const std::string& input) : Problem(input) {}
-    std::pair<bool, std::uint32_t> part1() override;
-    std::pair<bool, std::uint32_t> part2() override;
+    std::pair<bool, std::uint64_t> part1() override;
+    std::pair<bool, std::uint64_t> part2() override;
 };
